refactor: modularize game purchase flow hook
