+++ apps/web/components/game-purchase-flow/types.ts
@@ -1,0 +1,9 @@
+export type InvoiceFlowState =
+  | "idle"
+  | "creating"
+  | "polling"
+  | "paid"
+  | "expired"
+  | "error";
+
+export type CopyState = "idle" | "copied" | "error";
